22.2.24 10:50

+++ Ruwas/components/AllReusableComponets.js
@@ -322,7 +322,7 @@ export const ModifiedTextInput2 = (props) => {
       ]}
     >
       <TextInput
-        value={`${parseInt(value) + parseInt(dependentValue)}`}
+        value={`${parseFloat(value) + parseFloat(dependentValue)}`}
         multiline={true}
         onFocus={() => {
           setShow(true);
